feat(planfix_add_to_lead_task): move webhook handling
